backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
openai